+++ fivetran/data_source_external_logging.go
@@ -1,0 +1,10 @@
+package fivetran
+
+import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+func dataSourceExternalLogging() *schema.Resource {
+    return &schema.Resource{
+        ReadContext: resourceExternalLoggingRead,
+        Schema:      getExternalLoggingSchema(true),
+    }
+}